Changed layers.py APIs to be like Keras
