+++ dnet/layers/activations.py
@@ -1,0 +1,2 @@
+from jax.experimental.stax import Sigmoid as sigmoid
+from jax.experimental.stax import Relu as relu
